Introduced draw() method to all widgets
This method is called for the initial draw(). This is where the main
element may be modified, e.g. by adding classes or children to it.

This commit also removed the old widgetize() method, which is replaced
by a series of getter methods which may be used to change which elements
are used for styling or DOM events. The return value of these methods
may not change, i.e. it is not longer supported to change the the
element receiving events or classes etc. This has also never been used.