Components: handle global resize and visibility change events
This handling was previously part of the Root widget. Now all widgets
are able to handle resizing and visibility changes on their own. The
idea is that the top level widget in a tree will register global event
handlers and react to them. It can still for performance make sense to
have one top level container, however it is not required anymore.