Select: Add selected_entry option and redefine unselected state
This commit introduces a new option 'selected_entry' which contains
the selected entry currently selected. It also redefines the state
in which no entry is selected. In that situation the 'selected' option
will be -1, the 'value' option will be undefined and 'selected_entry'
will be null. Updated tests accordingly.

This commit also updates some part of the code to new code conventions.