Add optimized array diff utility functions
When updating e.g. equalizer bands option with a new set of eq bands, we
have to compare two arrays to decide which bands were removed and which
ones have been added.

When the equalizer is used from frameworks such as React, often
the same (but not necessarily identical) array is set as an option. To
optimize those situations we use optimized array diff functions in those
cases.