[testsuite]: Improve ergonimics of rendering area
Previously the rendering area was neither cleaned up after tests
properly nor was it properly hidden which could lead to weird scroll
behavior in the test suite.

We now remove all widgets from the rendering canvas after each test and
also make sure that the rendering area used does not influence the size
of the document.