Improve the handling of aria-label and labelledby
Previously, for widgets which have a label (i.e. Button, Fader, Meter,
ValueButton and ValueKnob), setting a label would remove any aria-label
and set the aria-labelledby to the ID of the label child widget.

The issue with this logic is that it makes it impossible to control
either aria-label or aria-labelledby from the outside. In particular, we
can end up in situations where aria-label was explicitly set but then
ends up as some default after the label was disabled.

This new logic means that aria-label is never influenced by visibility
of the label and aria-labelledby is only used if it is otherwise not set
by the user.