ranged: Refactor using renderers
This commit replaces the makeRanged function into a set of options,
renderers and events which have to be manually added to a widget.

The events will make sure to update the transformation and the
snap_module whenever any of the corresponding options changes. The
renderers will update the aria- attributes which correspond to min,
max and value.

This commit fixes a series of issues with the aria attributes:

- Allow applications to overwrite aria_valuemax, aria_valuenow,
  aria_valuemin manually.
- Format aria-valuemin and aria-valuemax using the value formatter.
- Update the aria-attributes correctly when the formatter changes.