Meter: Support array as gradient option
Add support for specifying a linear gradient using an array of objects
of the form { value: number; color: string; }. This syntax avoids issues
with the object syntax with values as keys, e.g. no need to parse
floating point values from string keys and the inability to specify the
same value twice.