utils/warning: Implemented a new mechanism
Currently the warning mechanism (using the aux-warn) class is
implemented using a timeout. It is set when a value it set which
is outside of the current min/max range.

The issue with this implementation is that it can spuriously trigger
the warning when min and max are updated at the same time as
the value itself.

This new mechanism relies on the rendering loop to set the
warning class (and install the timer). It will also remove the
timer and remove the warning class when min or max are
updated.