Widget: fix moving cursor in tabindexed widget's input elements
Explanation: we detect possible moving focus events on widgets to emit a proper event. This function also preventDefaults by default, although this breaks caret behavior on input elements (i.e. Value). this fix detects the type of the element to not preventDefault if it's an input element receiving the keyboard events.