Handle retargeted events
When events originate inside of a shadow dom and then
bubble across the shadow dom bounary, they get retargeted,
which means that their target property becomes the shadow dom
host element when emited outside of the shadow dom.

In some widgets this leads to issues because they track e.g. click
events outside by binding to the document click handler and
then checking if the target property.

We work around this by checking the composedPath() for events
with composed:true.